requirements.txt
streamlit
PyPDF2==3.0.1
google.generativeai
python-dotenv
streamlit_extras
